LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/wraps-successfully-expected.txt

item1 item2 item3

FAIL When the focus is set on the first item of a focusgroup that wraps, a backward arrow key press should move the focus to the last item within the focusgroup. assert_equals: expected Element node <span id="item3" tabindex="-1">item3</span> but got Element node <span id="item1" tabindex="0">item1</span>
